Enh: Move http_client to http/client
