+++ alignak/objects/schedulerlink.py
@@ -48,7 +48,7 @@ This module provide SchedulerLink and SchedulerLinks classes used to manage sche
 from alignak.objects.satellitelink import SatelliteLink, SatelliteLinks
 from alignak.property import BoolProp, IntegerProp, StringProp
 from alignak.log import logger
-from alignak.http_client import HTTPEXCEPTIONS
+from alignak.http.client import HTTPEXCEPTIONS
 
 
 class SchedulerLink(SatelliteLink):
