+++ test/etc/test_sslv3_disabled/scheduler-master.cfg
@@ -47,4 +47,5 @@ define scheduler {
     #satellitemap    poller-1=1.2.3.4:1772, reactionner-1=1.2.3.5:1773, ...
 
     use_ssl	     1
+    hard_ssl_name_check 0
 }
